+++ activejob/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Gemfile for new apps: upgrade redis-rb from ~> 3.0 to 4.0.
+
+    *Jeremy Daer*
+
 ## Rails 5.1.4 (September 07, 2017) ##
 
 *   No changes.